refactor script and add `perl` as an interpreter
